refactor: shared component index
