+++ src/shared/ui/react-hook-form/index.ts
@@ -1,0 +1,1 @@
+export { RHFTextField } from './RHFTextField';
