+++ src/shared/ui/switch-case/SwitchCase.tsx
