+++ src/shared/ui/switch-case/index.ts
@@ -1,0 +1,1 @@
+export { SwitchCase } from './SwitchCase';
